pyproject.toml
[project]
name = "pavlok-cli-agent"
version = "0.1.0"
description = "CLI agent experiment for Pavlok"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
